# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DaddyX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/create:
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new DaddyX event record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event detail with tokens and price history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/purchase:
    post:
      operationId: recordPurchase
      tags: [events]
      summary: Record a token purchase after on-chain confirmation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseBody"
      responses:
        "200":
          description: Purchase recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /events/{id}/report-revenue:
    post:
      operationId: reportRevenue
      tags: [events]
      summary: Admin - record revenue settlement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRevenueBody"
      responses:
        "200":
          description: Revenue reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/simulation:
    get:
      operationId: getSimulation
      tags: [events]
      summary: Get 20-point price curve simulation for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Simulation data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationPoint"
...
  /wallet/{wallet}/claim/{eventId}:
    post:
      operationId: claimRevenue
      tags: [events]
      summary: Fan claims revenue share for tokens held in a settled event
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revenue claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
...
  /events/{id}/holders:
    get:
      operationId: getTokenHolders
      tags: [events]
      summary: Get current token holders for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token holders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenHolder"
  /wallet/{wallet}/portfolio:
    get:
      operationId: getWalletPortfolio
      tags: [events]
      summary: Get tokens held by a wallet across all events
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet portfolio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioToken"
  /organizer/{wallet}/events:
    get:
      operationId: getOrganizerEvents
      tags: [events]
      summary: Get events for an organizer wallet
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organizer events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    Event:
      type: object
      properties:
        id:
          type: string
        eventConfigPda:
          type: string
        name:
          type: string
        venueName:
          type: string
        eventDate:
          type: string
        organizerWallet:
          type: string
        revenueShareBps:
          type: number
        initialPriceSol:
          type: number
        stepFactorBps:
          type: number
        payoutFactorBps:
          type: number
        tokenCount:
          type: number
        cancelled:
          type: boolean
        revenueReported:
          type: boolean
        settledRevenue:
          type: number
          nullable: true
        daddyxEnabled:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        tokensRemaining:
          type: number
        nextTokenPriceSol:
          type: number
        endDate:
          type: string
          nullable: true
      required:
        - id
        - name
        - venueName
        - eventDate
        - organizerWallet
        - revenueShareBps
        - initialPriceSol
        - stepFactorBps
        - payoutFactorBps
        - tokenCount
        - cancelled
        - revenueReported
        - daddyxEnabled
        - tokensRemaining
        - nextTokenPriceSol
        - createdAt
    Token:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        tokenId:
          type: number
        currentOwner:
          type: string
        currentPrice:
          type: number
        entryPrice:
          type: number
        revenueClaimed:
          type: boolean
        purchasedAt:
          type: string
      required:
        - id
        - eventId
        - tokenId
        - currentOwner
        - currentPrice
        - entryPrice
        - revenueClaimed
        - purchasedAt
    EventDetail:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/Event"
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/Token"
        priceHistory:
          type: array
          items:
            $ref: "#/components/schemas/PriceHistoryPoint"
        simulation:
          type: array
          items:
            $ref: "#/components/schemas/SimulationPoint"
      required: [event, tokens, priceHistory, simulation]
    PriceHistoryPoint:
      type: object
      properties:
        round:
          type: number
        price:
          type: number
        wallet:
          type: string
        timestamp:
          type: string
      required: [round, price, wallet, timestamp]
    SimulationPoint:
      type: object
      properties:
        n:
          type: number
        price:
          type: number
        organizerCumulative:
          type: number
        investorROI:
          type: number
      required: [n, price, organizerCumulative, investorROI]
    TokenHolder:
      type: object
      properties:
        tokenId:
          type: number
        wallet:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
        unrealisedRoiPct:
          type: number
      required: [tokenId, wallet, entryPrice, currentPrice, unrealisedRoiPct]
    PortfolioToken:
      type: object
      properties:
        tokenId:
          type: number
        eventId:
          type: string
        eventName:
          type: string
        eventDate:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
        unrealisedRoiPct:
          type: number
        revenueClaimed:
          type: boolean
        settled:
          type: boolean
        settledRevenue:
          type: number
          nullable: true
        revenueShareBps:
          type: number
        tokenCount:
          type: number
      required:
        - tokenId
        - eventId
        - eventName
        - eventDate
        - entryPrice
        - currentPrice
        - unrealisedRoiPct
        - revenueClaimed
        - settled
        - revenueShareBps
        - tokenCount
...
    CreateEventBody:
      type: object
      properties:
        eventConfigPda:
          type: string
        name:
          type: string
        venueName:
          type: string
        eventDate:
          type: string
        organizerWallet:
          type: string
        revenueShareBps:
          type: number
        initialPriceSol:
          type: number
        stepFactorBps:
          type: number
        payoutFactorBps:
          type: number
        tokenCount:
          type: number
        daddyxEnabled:
          type: boolean
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
      required:
        - name
        - venueName
        - eventDate
        - organizerWallet
        - revenueShareBps
        - initialPriceSol
        - stepFactorBps
        - payoutFactorBps
        - tokenCount
        - daddyxEnabled
    PurchaseBody:
      type: object
      properties:
        tokenId:
          type: number
        buyerWallet:
          type: string
        priceSol:
          type: number
        txSignature:
          type: string
      required: [tokenId, buyerWallet, priceSol]
    ReportRevenueBody:
      type: object
      properties:
        revenue:
          type: number
        adminWallet:
          type: string
      required: [revenue, adminWallet]
...
    ClaimResult:
      type: object
      properties:
        eventId:
          type: string
        wallet:
          type: string
        tokensClaimed:
          type: number
        revenueShareSol:
          type: number
        alreadyClaimed:
          type: boolean
      required: [eventId, wallet, tokensClaimed, revenueShareSol, alreadyClaimed]